+++ test/support/probe.rb
@@ -10,13 +10,16 @@ require "tmpdir"
 module Probe
   GEM = File.expand_path("../..", __dir__)
 
-  def boot(as: nil, **variables)
+  # `requests` are sent to the booted application in turn, each a `path` with the `method`, `input`
+  # and Rack `env` a client and the proxies in front of it would have given it.
+  def boot(as: nil, requests: [], **variables)
     Dir.mktmpdir do |directory|
       report = File.join(directory, "report.json")
       environment = {
         "RAILS_ENV" => "test",
         "DUMMY_DATABASE" => File.join(directory, "queue.sqlite3"),
-        "DUMMY_REPORT" => report
+        "DUMMY_REPORT" => report,
+        "DUMMY_REQUESTS" => JSON.generate(requests)
       }.merge(variables.transform_keys(&:to_s))
 
       _out, errors, status = Open3.capture3(environment, RbConfig.ruby, "-I#{GEM}/lib", "-I#{GEM}/test",